requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==2.2.4
langchain==0.1.8
flask==3.0.2
pypdf==4.0.2
python-dotenv==1.0.1
docx2txt==0.8
unstructured==0.12.4
python-pptx==0.6.23
openpyxl==3.1.2
google-generativeai==0.3.2
langchain-google-genai==0.0.9
chromadb==0.4.23
Spire.Doc==12.2.1
selenium==4.18.1
-e .